OAK-350: Unify PropertyState and CoreValue - change return type of count to int
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk/oak-core@1394158 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/jackrabbit/oak/plugins/memory/EmptyPropertyState.java
@@ -104,7 +104,7 @@ class EmptyPropertyState implements PropertyState {
     }
 
     @Override
-    public long count() {
+    public int count() {
         return 0;
     }
